# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/posts": {
      "post": {
        "operationId": "PostsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "posts"
        ]
      },
      "get": {
        "operationId": "PostsController_listAll",
        "parameters": [
          {
            "name": "subreddit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "posts"
        ]
      }
    },
    "/posts/{id}": {
      "get": {
        "operationId": "PostsController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "posts"
        ]
      },
      "patch": {
        "operationId": "PostsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "posts"
        ]
      },
      "delete": {
        "operationId": "PostsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "posts"
        ]
      }
    },
    "/posts/{id}/vote": {
      "post": {
        "operationId": "PostsController_vote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VotePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "posts"
        ]
      }
    },
    "/posts/{id}/mod/approve": {
      "post": {
        "operationId": "PostsController_modApprove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostModBaseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "posts"
        ]
      }
    },
    "/posts/{id}/mod/remove": {
      "post": {
        "operationId": "PostsController_modRemove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "posts"
        ]
      }
    },
    "/posts/{id}/verify": {
      "get": {
        "operationId": "PostsController_getVerify",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "posts"
        ]
      }
    },
    "/posts/{id}/audit-trail": {
      "get": {
        "operationId": "PostsController_getAudit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "posts"
        ]
      }
    },
    "/posts/{id}/mod/lock": {
      "post": {
        "operationId": "PostsController_modLock",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostModBaseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "posts"
        ]
      }
    },
    "/posts/{id}/mod/unlock": {
      "post": {
        "operationId": "PostsController_modUnlock",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostModBaseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "posts"
        ]
      }
    },
    "/posts/{id}/mod/pin": {
      "post": {
        "operationId": "PostsController_modPin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostModBaseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "posts"
        ]
      }
    },
    "/posts/{id}/mod/unpin": {
      "post": {
        "operationId": "PostsController_modUnpin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostModBaseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "posts"
        ]
      }
    },
    "/posts/{id}/mod/flag": {
      "post": {
        "operationId": "PostsController_modFlag",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostModBaseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "posts"
        ]
      }
    },
    "/posts/{id}/mod/unflag": {
      "post": {
        "operationId": "PostsController_modUnflag",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostModBaseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "posts"
        ]
      }
    },
    "/posts/{id}/comments": {
      "get": {
        "operationId": "PostsController_getComments",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "posts"
        ]
      }
    },
...
  },
  "info": {
    "title": "Jagoo Bahee API",
    "description": "API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "subredditId": {
            "type": "string",
            "example": "665b3f2a9c5a7d0012a1b234"
          },
          "authorId": {
            "type": "string",
            "example": "665b3f2a9c5a7d0012a1b235"
          },
          "title": {
            "type": "string",
            "example": "Check out my project"
          },
          "type": {
            "type": "string",
            "enum": [
              "text",
              "link",
              "image",
              "video",
              "crosspost"
            ]
          },
          "content": {
            "type": "string",
            "example": "Markdown content"
          },
          "url": {
            "type": "string",
            "example": "https://example.com"
          },
          "attachmentIds": {
            "example": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "crosspostId": {
            "type": "string",
            "example": "665b3f2a9c5a7d0012a1b240"
          },
          "flair": {
            "type": "string",
            "example": "Showcase"
          },
          "userSignature": {
            "type": "string",
            "example": "hex-signature"
          },
          "contentHash": {
            "type": "string",
            "example": "sha256:abcdef..."
          }
        },
        "required": [
          "subredditId",
          "authorId",
          "title",
          "type",
          "userSignature",
          "contentHash"
        ]
      },
      "UpdatePostDto": {
        "type": "object",
        "properties": {
          "authorId": {
            "type": "string",
            "example": "665b3f2a9c5a7d0012a1b235"
          },
          "title": {
            "type": "string",
            "example": "Updated title"
          },
          "content": {
            "type": "string",
            "example": "Updated markdown body"
          },
          "url": {
            "type": "string",
            "example": "https://example.com"
          },
          "attachmentIds": {
            "example": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "flair": {
            "type": "string",
            "example": "Showcase"
          }
        },
        "required": [
          "authorId"
        ]
      },
      "VotePostDto": {
        "type": "object",
        "properties": {
          "delta": {
            "type": "number",
            "enum": [
              -1,
              0,
              1
            ],
            "example": 1
          }
        },
        "required": [
          "delta"
        ]
      },
      "PostModBaseDto": {
        "type": "object",
        "properties": {
          "subredditId": {
            "type": "string",
            "example": "665b3f2a9c5a7d0012a1b210"
          },
          "moderatorId": {
            "type": "string",
            "example": "665b3f2a9c5a7d0012a1b200"
          }
        },
        "required": [
          "subredditId",
          "moderatorId"
        ]
      },
...
    }
  }
}